requirements.txt
alpaca-py>=0.10.0
pandas>=1.5.0
yfinance>=0.2.0
websockets>=10.0
pydantic>=2.0.0
environs>=9.0.0
structlog>=21.0.0
redis>=5.0.0
streamlit>=1.20.0
plotly>=5.0.0
requests>=2.25.0
alpaca-py
finnhub-python
tradingview-ta